// src/types/core.ts
export interface ASTNode {
  type: string;
  start: number;
  end: number;
  line: number;
  column: number;
}